Move the stb submodule to third-party directory
