+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "stb"]
-	path = stb
+[submodule "third-party/stb"]
+	path = third-party/stb
 	url = https://github.com/nothings/stb.git
