+++ CMakeLists.txt
@@ -298,7 +298,7 @@ find_package(libsodium REQUIRED)
 include_directories(${LIBSODIUM_INCLUDE_DIRS})
 set(LIBRARIES ${LIBRARIES} ${LIBSODIUM_LIBRARIES})
 
-include_directories(SYSTEM src/stb)
+include_directories(SYSTEM third-party/stb)
 
 # Protip, you must use a different directory for each build target...
 # -DMAKE_TOOLCHAIN_FILE has no effect unless the target directory is empty
