+++ src/stb.c
@@ -1,4 +1,4 @@
 #define STB_IMAGE_IMPLEMENTATION
-#include "../stb/stb_image.h"
+#include "../third-party/stb/stb_image.h"
 #define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "../stb/stb_image_write.h"
+#include "../third-party/stb/stb_image_write.h"
