+++ src/stb.h
@@ -1,8 +1,8 @@
 #ifndef STB_H
 #define STB_H
 
-#include "../stb/stb_image.h"
-#include "../stb/stb_image_write.h"
+#include "../third-party/stb/stb_image.h"
+#include "../third-party/stb/stb_image_write.h"
 
 // uTox can't find this function unless declared here.
 extern unsigned char *stbi_write_png_to_mem(unsigned char *pixels, int stride_bytes,
